% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_simplification.R
\name{network_simplification}
\alias{network_simplification}
\title{Network Simplification}
\usage{
network_simplification(links, nodes)
}
\arguments{
\item{links}{The links from the original network.}

\item{nodes}{The nodes from the original network.}
}
\value{
A list with links and nodes, as well as the proportion of communities included in the largest connected component and the modularity of the simplified network.
}
\description{
Extract the largest connected component of the original network and simplify it using the fast greedy modularity optimization algorithm into a network where each node is a population and each link is the number of individuals moving between populations.
}
\examples{
simplified_network <- network_simplification(links = original_network$links, nodes = original_network$nodes)
}
