% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_construction.R
\name{network_construction}
\alias{network_construction}
\title{Network Construction}
\usage{
network_construction(
  file,
  years,
  min_group_size = 1,
  min_link_size = 1,
  unix_time_origin = "1970-01-01"
)
}
\arguments{
\item{file}{The input file used for TILES, in the format specified by \url{https://github.com/GiulioRossetti/TILES}.}

\item{min_group_size}{The minimum group size to be included in the network (default is 1).}

\item{min_link_size}{The minimum link size to be included in the network (default is 1).}

\item{unix_time_origin}{The origin for the Unix times used in the input file for TILES (defaults to 1970-01-01).}

\item{year}{The years from which TILES snapshots were collected.}
}
\value{
A list with links and nodes.
}
\description{
Convert the output of the TILES dynamic community detection algorithm to a network where each node is a community and each link is the number of individuals moving between communities. Your workspace needs to include the unzipped graph and strong-communities output files from TILES.
}
\examples{
original_network <- network_construction(file = "edges.tsv", years = c(1970:1999))
}
